+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Cover `<LogViewer>` component with unit tests (UIDATIMP-710)
 * Cover `<FileItem>` component with unit tests (UIDATIMP-716)
 * Cover `<ProfileTree>` component with unit tests (UIDATIMP-718)
+* Cover `<Spinner>` component with unit tests (UIDATIMP-725)
 
 ### Bugs fixed:
 * Job profile fails because 1 action profile has 2 field mappings attached to it (UIDATIMP-931)